+++ scalr/resource_scalr_provider_configuration_test.go
@@ -771,6 +771,7 @@ resource "scalr_provider_configuration" "kubernetes" {
   name                   = "%s"
   account_id             = "%s"
   environments           = ["*"]
+  owners                 = []
   custom {
     provider_name = "kubernetes"
     argument {