+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.1] - 2024-09-27
+
 ### Fixed
 
 - Various false positive attribute drifts ([#344](https://github.com/Scalr/terraform-provider-scalr/pull/344))
@@ -883,7 +885,8 @@ Requires Scalr 8.0.1-beta.20200625 at least
 
 - Initial release.
 
-[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/Scalr/terraform-provider-scalr/compare/v2.1.1...HEAD
+[2.1.1]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.1.1
 [2.1.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.1.0
 [2.0.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v2.0.0
 [1.13.0]: https://github.com/Scalr/terraform-provider-scalr/releases/tag/v1.13.0